move path resolution to separate files
This commit mmoves uvwasi__resolve_path() and its helper
functions out of uvwasi.c. This reduces the size of uvwasi.c
by a few hundred lines, and should make it simpler to unit
test uvwasi__resolve_path().
